each header struct field is two bytes

+++ src/main.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone, Copy)] // TODO what other derives needed?
 struct DNS_Header {
-    id: u32,
-    flags: u32, // TODO bitflags?
-    num_questions: u32,
-    num_answers: u32,
-    num_authorities: u32,
-    num_additionals: u32,
+    id: u16,
+    flags: u16, // TODO bitflags?
+    num_questions: u16,
+    num_answers: u16,
+    num_authorities: u16,
+    num_additionals: u16,
 }
 
 #[derive(Debug, Clone)]
